--- README.md
@@ -1,3 +1,0 @@
-# final-18-moamoa
-
-멋쟁이사자처럼 7기 18조 프로젝트 test
