+++ moamoa/src/Components/Common/PostCard.jsx
@@ -1,10 +1,16 @@
 import React from 'react'
+import { Link } from 'react-router-dom';
 
 export default function PostCard(post) {
 
+  const baseUrl = `https://api.mandarin.weniv.co.kr/`
+
   const postprop = post.post;
-  const profileImgUrl =`https://api.mandarin.weniv.co.kr/${postprop.author.image}`;
-  const postImgUrl =`https://api.mandarin.weniv.co.kr/${postprop.image}`;
+  const profileImgUrl =`${baseUrl}${postprop.author.image}`;
+  const postImgUrl =`${baseUrl}${postprop.image}`;
+  const postDetailId = post.post.id;
+  console.log(postDetailId)
+  const postDetailUrl = `/post/${postDetailId}`
 
   const inputDate = postprop.createdAt
   const dateset = inputDate.split("").slice(0,10).join("");
@@ -18,15 +24,17 @@ export default function PostCard(post) {
       {post && (
         <li>
           <article>
-            {console.log(post.post)}
-            <img src={profileImgUrl} alt="사용자프로필"/>
-            <p>{postprop.author.username}</p>
-            <p>{postprop.author.accountname}</p>
-            <button>
-              케밥버튼
-            </button>
-            <p>{post.post.content}</p>
-            <img src={postImgUrl} alt="게시글 사진" />
+            <Link to={postDetailUrl}>
+              {console.log(post.post)}
+              <img src={profileImgUrl} alt="사용자프로필"/>
+              <p>{postprop.author.username}</p>
+              <p>{postprop.author.accountname}</p>
+              <button>
+                케밥버튼
+              </button>
+              <p>{post.post.content}</p>
+              <img src={postImgUrl} alt="게시글 사진" />
+            </Link>
             <p>{outputDate}</p>
             <p>{postprop.heartCount}</p>
             <p>{postprop.commentCount}</p>
