--- moamoa/src/Pages/PostDetail.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default function PostDetail() {
-  return (
-    <div>PostDetail</div>
-  )
-}
