+++ moamoa/src/Pages/Product/AddProduct.jsx
@@ -1,18 +1,18 @@
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useRecoilValue, useSetRecoilState } from 'recoil';
-import userToken from '../Recoil/UserToken';
+import userToken from '../../Recoil/UserToken';
 import axios from 'axios';
-import eventStateAtom from '../Recoil/EventState';
-import GoBack from '../Assets/icons/icon-arrow-left.svg';
+import GoBack from '../../Assets/icons/icon-arrow-left.svg';
+import eventStateAtom from '../../Recoil/EventState';
+const initialDate = new Date();
 
 const AddProduct = () => {
   const navigate = useNavigate();
   const [eventName, setEventName] = useState('');
   const [eventStartDate, setEventStartDate] = useState('');
   const [eventEndDate, setEventEndDate] = useState('');
-
-  const [eventPeriod, setEventPeriod] = useState(1);
+  const [eventPeriod, setEventPeriod] = useState(initialDate);
   const [imgSrc, setImgSrc] = useState(
     'https://cdn.visitkorea.or.kr/kfes/upload/contents/db/400_03e7c925-a8a5-4923-905c-e12586ec0a44_3.png',
   );
@@ -21,7 +21,6 @@ const AddProduct = () => {
 
   const setCategory = useSetRecoilState(eventStateAtom);
   const token = useRecoilValue(userToken);
-
   // API 요청-------------------------------------------
   const addEvent = async (imgSrc, eventName, eventPeriod, eventDetail) => {
     const baseUrl = 'https://api.mandarin.weniv.co.kr';
@@ -49,24 +48,35 @@ const AddProduct = () => {
         console.log(res.data);
       });
     } catch (err) {
+      //status 422
+      //에러 처리
       if (err.response) {
-        const { status, data } = err.response;
+        console.log(err);
+        // 요청이 이루어졌고 서버가 응답했을 경우
+        const { status, config, data } = err.response;
+
         if (status === 422) {
           console.log(data);
         }
+
         if (status === 404) {
           //404 이미지 출력
+          console.log(`${config.url} not found`);
         }
+
         if (status === 500) {
           console.log('Server error');
         }
       } else if (err.request) {
+        // 요청이 이루어졌으나 서버에서 응답이 없었을 경우
         console.log('Error', err.message);
       } else {
+        // 그 외 다른 에러
         console.log('Error', err.message);
       }
     }
   };
+  // API 요청--------------------------------------------
 
   const inputEventName = (e) => {
     setEventName(e.target.value);
@@ -79,7 +89,7 @@ const AddProduct = () => {
   const inputEventEndDate = (e) => {
     setEventEndDate(e.target.value);
   };
-  
+
   const inputEventDetail = (e) => {
     setEventDetail(e.target.value);
   };
@@ -87,8 +97,12 @@ const AddProduct = () => {
   const uploadImage = async (imageFile) => {
     const baseUrl = 'https://api.mandarin.weniv.co.kr/';
     const reqUrl = baseUrl + 'image/uploadfile';
+    //폼데이터 만들기
     const form = new FormData();
+    //폼데이터에 값 추가하기
+    //폼데이터.append("키","값");
     form.append('image', imageFile);
+    //폼바디에 넣어서 요청하기
     const res = await fetch(reqUrl, {
       method: 'POST',
       body: form,
@@ -99,6 +113,7 @@ const AddProduct = () => {
   };
 
   const handleChangeImage = (e) => {
+    //파일 가져오기
     const imageFile = e.target.files[0];
     uploadImage(imageFile);
   };
@@ -129,9 +144,7 @@ const AddProduct = () => {
 
   const submitProduct = (e) => {
     e.preventDefault();
-
     handlePeriod(eventStartDate, eventEndDate);
-
     addEvent(imgSrc, eventName, eventPeriod, eventDetail);
   };
 
@@ -160,6 +173,7 @@ const AddProduct = () => {
               onChange={handleChangeImage}
             ></input>
           </section>
+
           <section>
             <h2>카테고리</h2>
             <button type='button' onClick={handleEventTypeBtn} id='festival'>
@@ -169,6 +183,7 @@ const AddProduct = () => {
               체험
             </button>
           </section>
+
           <section>
             <label>행사명</label>
             <input
@@ -177,6 +192,7 @@ const AddProduct = () => {
               onChange={inputEventName}
               value={eventName}
             ></input>
+            {/* 확인 필요 */}
             <label htmlFor='event-period'>
               행사 기간
               <input
@@ -197,6 +213,7 @@ const AddProduct = () => {
               ></input>
             </label>
             {/* 확인 필요 */}
+
             <label>상세 설명</label>
             <textarea
               placeholder='행사 관련 정보를 자유롭게 기재해주세요.'
