+++ moamoa/src/Router/AppRoutes.js
@@ -9,7 +9,7 @@ import Join from '../Pages/Join/Join';
 import MyProfile from '../Pages/Profile/MyProfile';
 import YourProfile from '../Pages/Profile/YourProfile';
 import EditProfile from '../Pages/Profile/EditProfile';
-import FollowingList from '../Pages/Follow/FollowingList';
+//import FollowingList from '../Pages/Follow/FollowingList';
 import FollowerList from '../Pages/Follow/FollowerList';
 
 import ProductDetail from '../Pages/Product/ProductDetail';
@@ -21,11 +21,11 @@ import UploadPost from '../Pages/Post/UploadPost';
 import EditPost from '../Pages/Post/EditPost';
 import PostDetail from '../Pages/Post/PostDetail';
 
-import ChatRoom from '../Pages/Chat/ChatRoom';
-import ChatList from '../Pages/Chat/ChatList';
+//import ChatRoom from '../Pages/Chat/ChatRoom';
+//import ChatList from '../Pages/Chat/ChatList';
 
-import Error from '../Pages/Error';
-import Search from '../Pages/Search/Search';
+//import Error from '../Pages/Error';
+//import Search from '../Pages/Search/Search';
 
 export default function AppRoutes() {
   return (
@@ -74,10 +74,10 @@ export default function AppRoutes() {
       <Route path='/post/edit' element={<EditPost />} />
 
       {/* 게시글 상세 */}
-      <Route path='/post/:post_id' element={ProductDetail} />
+      <Route path='/post/:post_id' element={<PostDetail/>} />
 
-      <Route path='/user/login' element={<LoginPage />} />
-      <Route path='/user' element={<JoinPage />} />
+      {/* <Route path='/user/login' element={<LoginPage />} />
+      <Route path='/user' element={<JoinPage />} /> */}
     </Routes>
   );
 }
